Add more details on developer menu
It took me about an hour to find out how to invoke the developer menu in Genymotion, similar to  in the emulator on Mac. It may be obvious for those familiar with Genymotion, but I had to find  this article http://www.differential.com/insights/intro-to-debugging-react-native-ios-and-android/ and learn that in Genymotion that you can simply press a button or use  keyboard shortcut.

To avoid repetition throughout the docs on how to access this menu, I created a separate section for this.

Closes exponent/exponent-docs#31

fbshipit-source-id: d2f59b7