docs bugfixes, improvements, SEO opts (#836)
* [docs] upgrade all deps to latest (fixes #607)

* [docs] add noindex for unversioned URLs

* [docs] fix typo in expokit.md thanks to @ArmaizAdenwala

* [expo-docs] Document Sentry.enableInExpoDevelopment

Closes expo/expo-docs#112

* [docs] apply sentry doc updates to v19 as well

* [expo-docs] Improve ART docs

Closes expo/expo-docs#107

* [docs] update unversioned ART docs too

* [expo-docs] [docs] update Amplitude docs

Was reading through and saw that some links were broken, etc.

Closes expo/expo-docs#105

* [docs] update amplitude docs for unversioned

* [expo-docs] Add docs for new scheduling option intervalMs for local notif.

Documentation for https://github.com/expo/expo/pull/380

Closes expo/expo-docs#114

* [docs] update v19 push notifications docs to reflect correct API thanks to @timarney

* [expo-docs] Include iOS status bar text color info

Explain about using <StatusBar /> to keep the status bar text black as is the default in Expo (otherwise users will most likely be surprised when they see a difference between Expo and Standalone, and building/submitting to TestFlight isn'\''t the funnest of processes.)

Closes expo/expo-docs#102

* [docs] move note about default status bar color to building standalone apps guide

* [expo-docs] Update push-notifications.md

Improve documentation of push notifications:
- Add curl example
- Use non-deprecated api in example
- Mention api does not require auth

Note that it may be nice to mention weither or not expo tokens are stable unlike Apns/GCM ones that can be revoked by the provider. Not sure about that so I didn'\''t write it.

Closes expo/expo-docs#108

* [docs] apply push-notification changes to unversioned as well

fbshipit-source-id: a4931f2