Geocoding improvements (#902)
* Unify geocoding results between iOS/Android

On Android an empty array was returned when no results were found but on
iOS the promise was rejected with an error. Make both platforms return
an empty array.

* Add test cases for geocoding

* Add a note about permissions for geocoding on Android

On Android this permission is required, while on iOS it's not.

* Improve the error message for missing permissions

Trying to use geocoding without a location permission on Android
previously threw an error "Not authorized to use location services
Please set a Google API Key to use geocoding." because having
no permission and having no Google Play services were treated
the same way.

Distinguish between these errors and if the permission is missing,
rethrow the error "Not authorized to use location services".
If Google Play services aren't available, still fall back to using the
Google API and require the API key.

* [native-component-list] Add a ScrollView to GeocodingScreen

* Validate reverse geocoding params in JS

* Add unit tests for Location

* Make the unit tests more consistent in style

* Remove unnecessary mocks

* Partially revert changes in 27749b6

This mock is necessary for the tests to pass.

* Add geocoding documentation changes to unversioned folder

fbshipit-source-id: 78d2120