Fix some typos in the Glossary
By the way, the scroll position doesn'\''t always work well.

e.g. if you scroll down to the exp entry in the Glossary and click on the Read more link you get taken to guides/exp-cli.html, but instead of being positioned at the top of the page the scroll position is a little way down the page, as if something is remembering the scroll position in the Glossary and restoring it.

Also, clicking on an anchor (e.g. #yarn in the Package Manager entry) takes you to the right place, but clicking the browser'\''s Back button does not take you back to where you were before.

Closes expo/expo-docs#177

fbshipit-source-id: 74cccf6