Remove 'notification' mention from System Permission
I mentioned the wrong issue on the commit body.
The correct is this one: https://github.com/expo/expo/issues/1129

It seems that setting the 'infoPlist' for notification is not needed. (https://github.com/expo/expo/issues/1129)
This mention that I'm deleting seems to indicate that it is needed...
It took me a while to figure it out and it seems that mentioning this here will only create more confusion for others as it has created for me.

Thanks for your awesome work!!!

Closes expo/expo-docs#259

fbshipit-source-id: 1e56d34